requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
annotated-types==0.6.0
    # via pydantic
anyio==4.0.0
    # via httpx
asgiref==3.7.2
    # via django
certifi==2023.11.17
    # via
    #   httpcore
    #   httpx
django==4.2.7
    # via djangorestframework
djangorestframework==3.14.0
    # via -r requirements.in
exceptiongroup==1.1.3
    # via
    #   anyio
    #   pytest
h11==0.14.0
    # via httpcore
httpcore==1.0.2
    # via httpx
httpx==0.25.1
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   httpx
iniconfig==2.0.0
    # via pytest
packaging==23.2
    # via pytest
pluggy==1.3.0
    # via pytest
pydantic==2.5.1
    # via -r requirements.in
pydantic-core==2.14.3
    # via pydantic
pytest==7.4.3
    # via
    #   pytest-asyncio
    #   pytest-django
pytest-asyncio==0.21.1
    # via -r requirements.in
pytest-django==4.7.0
    # via -r requirements.in
pytz==2023.3.post1
    # via djangorestframework
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
sqlparse==0.4.4
    # via django
tomli==2.0.1
    # via pytest
typing-extensions==4.8.0
    # via
    #   asgiref
    #   pydantic
    #   pydantic-core
